main chunking done
