+++ freemocap_blender_addon/__init__.py
@@ -40,7 +40,7 @@ def unregister():
 
     try:
         print(f"Unregistering {__file__} as add-on")
-        from .blender_interface import BLENDER_USER_INTERFACE_CLASSES
+        from .blender_ui import BLENDER_USER_INTERFACE_CLASSES
         for cls in BLENDER_USER_INTERFACE_CLASSES:
             print(f"Unregistering class {cls.__name__}")
             bpy.utils.unregister_class(cls)
@@ -55,7 +55,7 @@ def register():
     import bpy
 
     print(f"Registering {__file__} as add-on")
-    from .blender_interface import BLENDER_USER_INTERFACE_CLASSES
+    from .blender_ui import BLENDER_USER_INTERFACE_CLASSES
     print(f"Registering classes {BLENDER_USER_INTERFACE_CLASSES}")
     for cls in BLENDER_USER_INTERFACE_CLASSES:
         print(f"Registering class {cls.__name__}")
@@ -76,7 +76,7 @@ def register():
 
     print("Registering property group FMC_ADAPTER_PROPERTIES")
 
-    from .blender_interface import FMC_ADAPTER_PROPERTIES
+    from freemocap_blender_addon.blender_ui import FMC_ADAPTER_PROPERTIES
     bpy.types.Scene.fmc_adapter_properties = bpy.props.PointerProperty(type=FMC_ADAPTER_PROPERTIES)
