+++ freemocap_blender_addon/blender_ui/operators/_add_rig.py
@@ -6,7 +6,7 @@ from freemocap_blender_addon.core_functions.empties.reorient_empties import reor
 from freemocap_blender_addon.core_functions.rig.add_rig import add_rig
 from bpy.types import Operator
 
-from ...blender_interface.operators._add_body_mesh import REORIENT_EMPTIES_EXECUTED
+from ...blender_ui.operators._add_body_mesh import REORIENT_EMPTIES_EXECUTED
 from ...freemocap_data_handler.operations.freemocap_empties_from_parent_object import \
     freemocap_empties_from_parent_object
