+++ freemocap_blender_addon/blender_ui/operators/_clear_scene.py
