+++ freemocap_blender_addon/blender_ui/operators/_reduce_shakiness.py
