+++ freemocap_blender_addon/blender_ui/properties/core_properties.py
@@ -1,5 +1,4 @@
 import bpy
-from bpy.props import StringProperty, BoolProperty
 
 from freemocap_blender_addon.freemocap_data_handler.utilities.load_data import get_test_recording_path
 
@@ -159,21 +158,4 @@ class FMC_ADAPTER_PROPERTIES(bpy.types.PropertyGroup):
         description='Mode (source) for adding the mesh to the rig',
         items=[('custom', 'custom', ''),
                ('file', 'file', '')]
-    )
-
-
-class FMC_VIDEO_EXPORT_PROPERTIES(bpy.types.PropertyGroup):
-    export_profile: bpy.props.EnumProperty(
-        name='',
-        description='Profile of the export video',
-        items=[('debug', 'Debug', ''),
-               ('showcase', 'Showcase', ''),
-               ('scientific', 'Scientific', ''),
-               ],
-    )
-
-    ground_contact_threshold: bpy.props.FloatProperty(
-        name='',
-        default=0.05,
-        description='Ground contact threshold (m)'
     )
